created files *.jsp & little fix

+++ src/main/java/com/bookmanager/dao/BookDao.java
@@ -12,9 +12,9 @@ public interface BookDao {
 
     void updateBook(Book book);
 
-    void removeBook(Long id);
+    void removeBook(int id);
 
-    Book getBookById(Long id);
+    Book getBookById(int id);
 
     List<Book> listBook();
 }
